test: increase size estimate fuzz factor
Increases error tolerance for size estimation vs reported size. The
error bar was 100mb, and now it is 115mb. Given that the test needs to
pass even when things like an unpredictable amount of data going into
the controller log and snapshots, the fuzz factor was pretty high. Given
new features are added to controller etc.. it's not totally unreasonable
to think they are just a little bit bigger on average.

Fixes: https://github.com/redpanda-data/redpanda/issues/14320

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>