rptest: Fix race condition in the test suite
The tiered-storage e2e tests are using BucketView which is scanning
periodically to do some validations. It is possible for the BucketView
to scan the bucket and get the object listing right before one of the
segments is deleted. In this case the BucketView will throw exception
and the test will log error. This commit fixes this by handling NoSuchKey
error.