iceberg: add logical data types
Adds initial logical types[1] that will be used to represent an Iceberg
schema. These include "primitives" (basic types like int, float, string,
etc) as well as complex types (types that are composed of other types,
like list, map, struct). Complex types are represented in Iceberg as
composed of "nested fields" which are types that are themselves composed
of other types (primitives or complex types).

This PR introduces the recursive definition of these types, with a basic
equality operator and ostream operator (note, this is for JSON
serialization, that will come in a following commit).

The implementation of these types is similar in style to the Iceberg
Rust library, though we are using std::variants instead of Rust enums,
since C++ enums aren't expressive enough to express complex nested enums
with members.