iceberg: introduce json utils
Adds some helpers for operating on the json::Value class, that will be
useful in building JSON parsing for Iceberg metadata types.

Long term it may make sense to move these to some general purpose
module, but since this is only used in Iceberg for now, and to avoid
distractions, the utilities are just added to the Iceberg module.