cst/ducktape: Change assertion for deleted chunks
Because chunk readers keep open the handle to a chunk file even if the
file itself is deleted, during the test_read_chunks it is possible that
the sequential consumer is entirely served by these open file handles
while all chunk files in the directory have been removed. This results
in the assertion that some chunks must have been removed during
sequential consumer to fail.

As the count of chunks deleted is not important as an assertion and the
purpose of deleting chunks randomly was to exercise failed
materialization code path which should happen during the random consumer
run, the assertion is changed to only test once for both seq and random
consumers.

Another alternative would be a more fine grained check, eg assert if
some chunks are removed only if some files are written to the cache
during a consumer run, but such checks do not buy extra value.