tests/controller_snapshot: fix no-op controller leadership transfers
Sometimes new controller leader won't be elected until a few seconds
after a controller restart. This can confuse subsequent leadership
transfers that will try to move controller leadership to the old
controller and falsely think that it will be a no-op. Wait for the new
leader after the joiner node is restarted to fix that.

Fixes https://github.com/redpanda-data/redpanda/issues/20572