tests/maintenance_test: increase the number of partitions
Previously when this test was running in CI there were not a lot of partitions (12)
and the leadership balancer could mute them all after several ticks (it mutes the
partitions that it moves). After that it cannot take any further action for some time
and the node that comes out of the maintenance mode won't acquire any new leaders.

Increase the number of partitions in the test to avoid that.

Fixes: https://github.com/redpanda-data/redpanda/issues/13650
(cherry picked from commit a6456189adb1e1343be2fd7bdb99bb6995b829d9)