sr/json: drop support for draft 5
This is needed because the next commit introduces jsoncons for
validating json schemas against their draft's metaschema and jsoncons
does not have support for draft 5.

Other schema registries do not support draft 5 either, so it is unlikely
that we would have customers that need to use draft 5 explicitly (and
can't just use the largely equivalent draft 4).