iceberg: add missing field to manifest_entry avro
Adds the file_sequence_number to the manifest_entry schema, as indicated by the
spec[1].

It's unclear why the schema didn't have them in the first place (the
schema was taken from DuckDB), but it's clearly defined upstream[2] and
is important for efficiently writing metadata[3].

There is another field, distinct_counts, in the spec that is not in the
manifest_entry schema. Interestingly enough, it doesn't appear to be
defined at all in upstream Java code[4] (it is defined in iceberg-go,
but not iceberg-python or iceberg-rust), so I'm leaving it out.

[1] https://iceberg.apache.org/spec/#manifests
[2] https://github.com/apache/iceberg/blob/319f29ea860e42e7cc21cda8c05d882134e6431f/core/src/main/java/org/apache/iceberg/ManifestEntry.java#L48-L49
[3] https://iceberg.apache.org/spec/#manifest-entry-fields
[4] https://github.com/apache/iceberg/blob/319f29ea860e42e7cc21cda8c05d882134e6431f/api/src/main/java/org/apache/iceberg/DataFile.java#L37-L105