iceberg: pull ownership of iobuf out of avro_iobuf_ostream
When using a DataFile{Writer,Reader}, we'll need to be able to pass the
buffer used by the writer. This is difficult to do with the current
interface of avro_iobuf_ostream, which is std::moved into the writer.

So, this moves the buffer outside of the ostream, which makes it easier
for callers to manage the iobuf.

Note, it doesn't appear the trimming I was doing on release() was needed
after all.