cloud_storage: add `topic_namespace_override` to `remote_path_provider`
The `remote_path_provider` will use a provided `topic_namespace_override`
in the lieu of provided topic namespace specifiers provided to its path
generation functions in the future.

This commit adds the `_topic_namespace_override` to the class, and amends
call sites to the constructor throughout the code using `std::nullopt` as a
placeholder.

These will be individually fixed to use the `topic_namespace_override` from
`topic_properties` in future commits.