rpk/transform/build: fix js sdk tooling
Mac builds dynamically link the executable instead of static linking on
linux. Make sure the mac builds include the required dylib (in the right
directories too so rpath fiddling is not needed).

I packaged a new toolchain using `package_toolchain.py`, uploaded that
to the tinygo repo (still need a better place to upload that stuff), and
fixed the paths now that they changed in the tarball.