c/tx_gateway: fixed incorrect mapping of error code in add to txn
Changed the way how the error returned from update transaction in
`do_add_parititons_to_txn` is handled. Previously every error in state
update resulted in `invalid_txn_state` error. This not the case for most
of the errors as client should simply retry the request. Beginning
transaction on data and group partitions is idempotent so it can be
safely done once again when handling request retry.

Fixes: CORE-5698

Signed-off-by: Michał Maślanka <michal@redpanda.com>