x/sdk/js: Improved error reporting
Adds `is_error` to distinguish between JavaScript errors
and exceptions.

Also adds a stack trace to the debug string for exceptions and
errors both. Note that traversing the stack trace is a recursive
operation, but generally we're going to be hitting this code when
an exception escapes a transform function, killing the VM.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>