cst/config: Disable trim carryover
The trim carryover feature remembers items to be trimmed to quickly free
up space on trim without going to disk.

With access tracker based trim, carryover can remove files, the updates
to the inner data structure of access tracker can be pending and the
tracker can try to delete these files again. This can potentially result
in error logs due to double-delete of files.

The carryover limit is set to 0 so that the list does not carry any
items. A user can still set the list to some value and enable the
feature, but this can result in error logs.

(cherry picked from commit 71210aa3f6b02569efaea493701fa57a2c56dc6f)