tx/gateway: set outcome result when sync in end_txn failed
If transactional coordinator sync fails the returned result should be
forwarded to the client. This way a client can correctly handle the
error and retry a request if necessary. With the current code the
`end_txn` outcome promise was not set when sync failed, leading to
Redpanda returning an `unknown_server_error` to client.

Signed-off-by: Michał Maślanka <michal@redpanda.com>