cloud storage: Fix deleted_count in clean_up_at_start
`clean_up_at_start` deletes files that are not in the access time
tracker. While doing so it keeps a count of the total bytes deleted as
well as the number of files deleted. Before this change, the number of
files deleted was tracked incorrectly, causing the cache to start with
incorrect values for `_current_cache_size` and `_current_cache_objects`.