cloud_storage: add `topic_mount_handler`
The `topic_mount_handler` class is a thin wrapper around `topic_mount_manifest`
operations in cloud storage.

Most notably, it provides the `mount_topic()` and `unmount_topic()` functions,
which construct a `topic_mount_manifest` and `remote_path_provider`, and
perform the necessary cloud storage checks/operations for mounting a topic
from a provided `topic_configuration`, and returns the outcome in form of a
`topic_mount/unmount_result` enum.

The cloud storage operations used for mounting are currently non-atomic,
with individual calls used to 1. check for the existence of a
`topic_mount_manifest` and to 2. delete it. This could lead to race
conditions involving multiple clusters and topic mounting in the future, as
`DeleteObject` requests will return a success outcome if a key to be deleted
is not found in cloud storage.

(cherry picked from commit bbd86d4767de76839922f89747b7c28c7e9fc572)