cloud_storage: add `topic_mount_manifest`
Adds the `topic_mount_manifest` definition. This class is constructed
from a `remote_label` (intended to be the "source" label- the label
for the cluster which originally created the topic of interest) and a
`topic_namespace`.

The `topic_mount_manifest` is meant to indicate ownership of an
existing topic in case of migrations- if it exists in tiered storage,
then there is currently no cluster producing or consuming to that topic,
and a cluster may safely take ownership of it by deleting the manifest
and mounting the topic.

The manifest path is currently of the form:

`migration/<source_label>/<namespace>/<topic>`

(cherry picked from commit f581eeecec21839e4530d13e6fb39893d031d76c)