cluster: Reset cloud storage metric...
... in the partition::remove_persistent state. The probe has access to
the global state so it's not safe to keep it after the partition is
stopped and `remove_persistent_state` method is invoked. In cases when
the partition is recreated with different revision id the 'double metric
registration' error could be triggered.

The double metric registration exception prevents normal parittion
removal and subsequent partition creation may trigger assertion.