iceberg: refactor Avro ostream to avoid iobuf
While ultimately we want an iobuf from the ostream, it currently
requires pointer stability of the underlying fragments, which isn't an
explicit contract of the iobuf. Instead, this tweaks the class to use a
container of temporary_buffers directly, and expects callers to pass the
buffers to an iobuf.

There was some discussion on making iobuf smarter and providing an
interface that guaranteed pointer stability of buffers. This seems like
it'd be a heavy-handed requirement to impose on iobuf to enfroce.
Putting the onus on callers seems like a quick, good-enough approach.