k/client: Retry connection on DNS error
During a rolling upgrade in cloud, it was observed that RP's kafka
client would attempt to connect to the 'reserve' node after it was
decomissioned.  This was because the error code (C-Ares ENOTFOUND) was
not treated as a retriable error.  This change checks for the above
error code when attempting to connect to a broker and if it is
encountered, treats it as a retriable error.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit d79ab902239e382646bf8a81ccb14af56b71d524)