rpk: change move partition to use new endpoint
Redpanda clusters are switching to a node-local
core assignments, which means that the API for
dispatching cross-core partition moves will be
different.

This is guarded behind the new feature:
node_local_core_assignment. Those clusters with
the feature enabled will issue the core movements
independently from the node changed.

If it's not enabled/present, we still do the old
API calls.