cluster: move archival into cluster directory
Today archival is built directly into the cluster library because there
are dependencies in both directions. Despite being built into the
cluster library, all of the archival files are physically located in a
separate directory. CMake allows this, but Bazel flags this as a
violation of package boundaries. It can be worked around, but it doesn't
seem to be worth it.

The goal for Redpanda is to fully split out the archival library from
cluster, it's merely a matter of someone taking on the refactoring
project.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>