cst/inv: Fix aws ops prefix handling
The aws ops methods assumed that if a prefix is supplied to listv2, the
result contains common-prefixes which start after the prefix. The result
actually contains the full path from the bucket root.

For example if prefix=foo/bar and there is a directory called 1 within
foo/bar then the common-prefix in list-result is foo/bar/1/ and not 1/.
This mistake is fixed in this change, and the associated tests are also
fixed.