tests: increased tolerance in fast addition tests
In order to make the test stable increased a tolerance on the size of
data on a node that joins the cluster. The slight increase doesn't
influence the test correctness but makes it less vulnerable to small
changes in size of partition metadata.

Fixes: #20224

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 7674dc04420d7fd982252a459f4c106b56baca2a)