cloud_storage_clients: fix `abs_parse_impl::handle_start_element`
We ran into a case in which the `xml_sax_parser` would throw an exception
when processing a ABS `List Blobs` response like the following:

<?xml version="1.0" encoding="utf-8"?>
<EnumerationResults ServiceEndpoint="https://blob.core.windows.net/" ContainerName="container">
  <Prefix>cluster_metadata/bb7527f1-3227-4d55-86da-c133ec955ea9/manifests/
  </Prefix>
  <Delimiter>/</Delimiter>
  <Blobs>
    <BlobPrefix>
      <Name>cluster_metadata/bb7527f1-3227-4d55-86da-c133ec955ea9/manifests/2/
      </Name>
      <Properties>
	<Creation-Time>Thu, 25 Jul 2024 14:07:26 GMT
	</Creation-Time>
      </Properties>
    </BlobPrefix>
  </Blobs>
</EnumerationResults>

The issue here being that no `<Blob>...</Blob>` element is returned, and when
processing the `<BlobPrefix>` element an exception is triggered.

This commit corrects the parser code to anticipate this potential corner case.