cloud_storage: add `error_outcome::operation_not_supported`
Adds a value to the `cloud_storage::error_outcome` enum for handling cases
where the response to a request made had some flavor of `OperationNotSupported`
code.

This will be used immediately to handle a case for the `abs_client` where
`Delete Blob` requests made on directories results in a
`OperationNotSupportedOnDirectory` response.