archival: fix `purger::collect_manifest_paths()`
Before, the `purger` would push back what it assumed was the spillover manifest
file by default to its list of `collected_manifests`.

In the case of ABS, `_api.list_objects` might actually return the directory
itself as a `Blob`. This would lead to the `purger` attempting to download
the directory as if it were a manifest, which would always fail.

This would completely block the `purger` from progressing and deleting
other partitions in the deleted topic, as it would retry the same doomed manifest
download.

Correct the logic in `collect_manifest_paths()` by checking the path for
`manifest.bin`, which should be contained within the spillover filename
(e.g `.../5_21/manifest.bin.10.11.0.1.999.1000`).

(cherry picked from commit e3574ffbbfde3eb09b12786d6270a2fba4c85937)