kafka: move klog into the protocol library
It really should be move out of the protoco library into the server
library where it is used. However, the batch decoders in the protocol
layer use it, and its a bit of a challenge to remove. An example of the
difficulty is that the generated kafka protocol encode/decode routines
invoke the logger in many levels of indirection. One way to address this
might be to throw an exception with the context, as all cases appear to
shutdown the connection whenever a problem occurs that triggers the log
message. But that will require more careful analysis.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>