k/group: fixed mapping of transactional control batch replication errors
When replication of transactional control batch fails group coordinator
steps down and return an error to transaction coordinator. Fixed the
incorrect mapping of replication errors which may lead to returning an
`unknown_server_error` to client.

With the new mapping policy the error handling in consistent for all the
replication batches. All the errors are translated to
`not_coordinator_error`. The errors indicating uncertain outcome like
shutting down and timeout are translated to timeout to let the client
know that the outcome of operation must be verified.

Signed-off-by: Michał Maślanka <michal@redpanda.com>