WIP cluster: rename read_replica_bucket as bucket_override
This repurposes the read_replica_bucket to instead be a bucket_override.
Whereas before, the property would only ever be set to indicate being a
read replica, moving forward we would like the ability to set it on
non-read replicas. To that end, this commit changes its name and tweaks
its semantics to not be tightly coupled with is_read_replica.

WIP:
- bucket override is duplicated in archival_metadata_stm and
  cluster::partition
- need to decide what to do about cache collisions
- needs tests