`rptest`: fix `full_disk_test`
The `log_eviction_stm` will only be triggered by the `log_manager`'s garbage collection
if a `log` has more than one `segment` in it.

This test previously produced to a topic based on a total number of bytes
(`produce_until_bytes`), which would sometimes result in a partition having
a log with only one segment. This would not be deemed eligible for garbage
collection, and therefore the test would timeout, since the on-disk usage
would never be reduced passed the threshold of 1MB.

Alter the test to produce at least 2 segments (`produce_until_segments`),
so that there is no possibility of a log being ineligible for garbage collection.