kafka: rename list_offsets header to list_offset
The kafka message `list_offsets` uses a non-plural name in the kafka
protocol specification, but our headers use a plural name. This commit
changes the header file names to match the code generated names so that
we do not have to special case the message in the bazel macros.

Newer versions of kafka message specifications have changed the name to
be plural. So the next time list_offsets message version is bumped and
we bring in an updated specification then we'll need to align the file
names again.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>