kafka: use valid versions instead of server supported
The protocol test used an api version range defined by the range
supported by the redpanda server. This not only created a dependency
on the server from the protocol library, but it limited the scope of
testing since the server doesn't support the entire valid range of every
message.

This commit removes the dependency, and expands testing by using the
valid range of the message as defined in its specification.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>