`admin`: add `key` to `get_cluster_config` json
`rpk cluster config get` makes a request for a single property from
the cluster.

Previously, this would return the entire `configuration` json to `rpk`,
which would pick the key out for the sole property it wanted.

Add `key` as an optional query parameter to the `get_cluster_config`
HTTP route, in order to support returning just a single property from a cluster
config request.

This will also support using property aliases with `rpk cluster config get`
in the future.