raft: introduced buffered protocol
Introduced a `raft::buffered_protocol`.

Buffered protocol is a wrapper around `raft::consensus_client_protocol`
tha the main purpose is to control the flow of requests from all the
raft groups instantiated on one shard to the same remote node. Instead
of tracking the requests separately per each raft group the buffered
protocol tracks and buffer all of them targeting the same node. The size
of the buffer and number of inflight append entries requests is tunable
with the cluster configuration properties.

Using a single buffer per shard for the same follower allows to better
amortize spikes in the follower disk latency without propagating it to
the end user.

Signed-off-by: Michał Maślanka <michal@redpanda.com>