bazel: remove layering check exclusion
The layering check in Bazel is a clang feature that checks that included
headers have directly listed dependencies, rather than relyin go
transitive dependencies.

For some of our dependencies we use cmake/autotools and wrap these to
avoid creating native bazel builds (yet) for these dependencies. The
tools called foreign_wrapper_cc are compatible with the layering checks,
so we had excluded many redpanda libraries from the checks.

This PR works around the limitation by adding a layer of indirection. We
create a sort of pass through library for each external dependency that
has this issue and disable layering checks for this library. In turn the
pass through library contains headers that mirror the real headers and
include them.

We then include the pass through library instead of the direct
dependency, which allows layering checks to be enabled for all redpanda
builds while constraining the relaxation of the rule as much as
possible.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>