storage: fix condition in `disk_log_impl::have_segments_to_evict()`
Currently, we check that there is more than one segment in a `disk_log_impl`'s
`segment_set` for determining it as having segments for eviction.

This gets us into a situation in which we can evict all segments in a log if
we were to have `_segs.size() > 1`, but none if `_segs.size() == 1`.

This is an overly restrictive condition, and is corrected here.

Also, add a new test to `log_retention_tests.cc`, `retention_test_size_with_one_segment`
to ensure test coverage for this new condition.

Also, adjust `retention_ms` in `test_bogus_timestamps`.
A low `retention_ms` value in this test case quickly evicts the
lone active segment with a valid timestamp, leading it to timeout when checking
for `"Adjusting retention timestamp.*to max valid record"` in the log
monitor in the case of `mixed_timestamps=True, use_broker_timestamps=False`.
Raise its value to `10,000` (`10s`) to avoid this issue.

(cherry picked from commit 2d007c6bdb4bf0e3ba1ab7319c814ada91c0e88d)