container: fix frag_vector<bool>
Previously reading data from frag_vector returned a reference (T& and
const T&). However std::vector<bool> does not return references from
methods like `at` and `operator[]`. So the following code FAILS to
compile:

```
std::vector<bool> vec;
vec.push_back(false);
bool& elem = vec[0]; // ERROR: cannot take a reference to a temporary
```

In order to fix this we return copies instead of references for
`fragmented_vector<bool>`. This is fine because the copies are trivial,
and there are no instances of this in the tree at the moment (by
definition because it doesn't compile). The only way to keep everything
as references is to remove `std::vector` as the inner container to
fragmented_vector.