c/m_manager: pass offset of cluster bootstrap command to members manager
When `members_manager` is updated with the initial state the update the
update must be ordered after raft configurations that may be preceding it.
This way the order of updates is preserved and no connection updates is lost.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 0d4f825d8be459acb2a84a200269a32c959cbf0b)