c/members_table: fixed incorrect validation in set_initial_nodes
Previously an invalid assertion in the `members_table` was actually
passing as `controller` raft was initialized before any of the features
were active. This made `controller` configuration to always contain a
broker and that configuration was applied to the `members_table` before
any of the other updates. Fixed the assertion and change the behavior to
always prioritize the `set_initial_nodes` and not the `controller` Raft
configuration.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 8f43d80c77f7b8f3efa33b7e20e7c4b538c490ec)