dt/acl: add tests for validating kafka ACLs
This adds a ducktape tests that validates that the correct ACLs are
being applied to certain kafka endpoints.

Specifically the tests are added are for the describe and alter client
quotas endpoints, but the intent of this change is to make it easy to
add new tests to validate the ACL-requirements of newly added endpoints.