tests/si_utils: introduce small tolerance for the lower bound of size
When topic is recovered there may be a slight variability in the size of
the last recovered segment. This variability made the recovery test to
be flaky. Added a small 5 percent tolerance for the lower bound of
recovered segments size.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 86d2a2c87ef2411277cbbac2abcf70e12cbce36a)