cst/inv: Read max hash size during parse from config
The inventory service will write data to disk during inventory report
parsing when a certain memory size is reached. Previously this size was
hardcoded to 128 KiB. This can result in many files being written to
disk for large inventory reports. This size is now a config setting and
defaults to 128 MiB.