r/stm_manager: do not warn when some of the stms did not make progress
When a partition is created without the batch cache, the write may not
be immediately visible for the reader. In this case the
`raft::state_machine_manager` identifies it as an issue with the state
machine not being able to make progress. Added a differentiation of the
log level based on the cache configuration of the partition. For the
partitions that uses the cache the `WARN` log level is kept to make it
easy to recognize possibly stuck state machines.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 53e9fb4ca492572d1aa8f7c8e8f2f22d9fad6fe6)