net: make rejected connection naming consistent
We can reject connections because we hit two types of limits: the
open connection limit or the rate limit (in the rate limit case
we do try to wait a short period to see if accepting the connection
after the wait would meet the rate, but if not we reject).

Existing naming of variables, log lines etc was inconsistent using
"declined" for the second case and generally not being clear about
the two cases.

Bring this into line by using "rejected" as the standard terminology
and then "open limit" and "rate limit" being the two reasons.

Log the addr in the same way too, including the port, in either case.

(cherry picked from commit e55a2bec2d1f898c031a84c3bcf247baba5c3a0b)