cst: Use ntp hashes in scrubber/anomaly detector
When checking for segments within a manifest, the anomaly detector first
tries to load up ntp hashes which could have been placed there by
inventory service.

Once loaded, this data is used as following:

* for each segment path, first check the inv. data
* if path exists move on. no operation is consumed
* if path missing or collision, make http call
* if path still missing, mark the segment as missing in anomalies

The hash data set is not checked for manifests as these have to be
downloaded for the scrub.

If the hash data set was not loaded, then the segment-exists checks are
completely skipped. This makes sure that we do not make a lot of HEAD
requests if the data set is missing. In this case the segment existence
check is skipped.

There are two caveats here, if scrub is enabled but inv-based scrub is
disabled, we always make HTTP calls, because this config is deemed to be
explicit intent that such calls should be made. Second is that a flag to
force HTTP calls is provided in the anomaly-detector::run method. If set
then HTTP calls are always made even if data set is missing. This is for
the benefit of the topic recovery validator, where we may want to always
make HTTP calls even if there is no inv. data.