c/health: validate target_node_id when collecting health report
The health report is used to determine if a cluster node is online and
available. When a node id changes but the RPC endpoint does not change
the requester may incorrectly assume that the node with the previous
node_id but the same endpoint is still operational. Added validation of
the node that the request was sent to before collecting the health
report. This way a sender will have correct information about the node
availability as only the request targeted to the node with the correct
node id will be replied with success.

Fixes: CORE-5766

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 90eafa83727cdd2f8ab3bdd7db4c65b2c3c50cbc)