archival: Fix upload policy for compacted topics
The archival_policy::find_segment method has a bug which causes it to
skip compacted segments unconditionally. The code checks if the segment
that it found is compacted and if this is the case it advances the
iterator to the next segment. This operation causes ntp_archiver to skip
one segment. The uploaded segment doesn't line up with the previous one
so the validation error is triggered.

The commit also implements handling of gaps in a safe way. We're not
supposed to have gaps in the log (even in compacted logs) so the
implementation has linear complexity. It's not supposed to be used and
is good enough to hadnle unexpected unique glitch.

(cherry picked from commit de013217e1887ca87330144695bfd6cb94dbb589)