kafka/group: don't replicate empty batch on offset commit
Redpanda generally does not permit empty batches to make their way to
the log. This is problematic in case of an empty offset commit requests.

Kafka returns immediately for empty requests[1], so this tweaks the
return store_offsets() method to do the same. I considered putting this
higher up in the Kafka layer, but there didn't seem to be a specialized
handler for offsets commits where such an early return would be natural.

[1] https://github.com/apache/kafka/blob/98cdf9717049e87ba34bb5161276577fcb8bd1c4/core/src/main/scala/kafka/server/KafkaApis.scala#L504-L505