`cluster`: update `serde::version<>` for `get_status_response`
This struct was erroneously altered by adding a new field without bumping
the `serde::version<>`. This struct also lacks a `serde_fields()` implementation,
so adding the new field anywhere but the end of the structure is an issue
for serialization/deserialization.

Update it by moving the field to the end of the struct, and bumping the
`serde::version<>` to 1.

(cherry picked from commit fb3fd6df7cc63a5695d10b8c75eae96c6d06059e)