partition-allocator: default shard 0 parts to 0
We have the topic_partitions_reserve_shard0 property which makes the
partition allocator have as if there are that many partitions
allocated to shard0, beyond the amount actually allocated. The idea is
to slightly bias the partition related work on shard 0 in order to
account for extra duties it has to do.

This is by default set to 2. This change sets it to 0 instead. In
general the default value of 2 does not do a good job of accounting
for the bias: it will either be much too high or much too low in most
cases. I.e., if every shard has 500 partitions, putting "only" 498
makes no difference, despite that shard 0 might be doing significantly
more work in this case. On the other hand, if someone takes our
"recommendation" of using 1 or 2 shards per partition for maximum tput,
they might be surprised to find that shard 0 get no partitions at all,
since it has these two reserved slots. In this case shard 0 will
be seriously underutilized.

Especially due to this second effect, which does come up in practice,
it seems like 0 is the best default value for this property. The idea
of choosing a fixed value to reduce the load on shard 0 seems appealing
on the surface but does not seem to work out in practice.

CORE-6852