k/group_probe: sanitize group names in metrics
The group name is returned as-is in the output of `/metrics` and
`/public_metric`. This means that the metrics response can be malformed
if the group name containes a `"` character.

This fixes this problem by sanitising the generated metric labels.

As a follow-up, we should also validate consumer group names when the
group is created, matching Kafka's validation logic.