`cluster`: fix `partition::may_read_from_cloud()` for RRR
`may_read_from_cloud()` was modified with a sanity check with
`remote_fetch_enabled()` to assert that `remote.read` is enabled for a
topic before serving a timequery from cloud storage. However, it failed
to consider the case of a remote read replica, which should always be
able to read from cloud.

Fix the condition by adding an `or` statement to `may_read_from_cloud()`
which checks `is_read_replica_mode_enabled()`.