bazel: introduce implementation deps
implementation deps hide the headers to transitive targets, which is
really handy for keeping private headers private. The layering check can
be used in some cases to prevent transitive dependencies leaking when
coupled with bazel visibility, but headers can still accidently leak if
one is not careful. Implementation deps is a much better solution.