`cloud_storage_clients`: add `infer_backend_from_uri()`
In the case of S3-compatible providers, we determine the backend
through the access-point/uri if we are already unable to determine
the backend from credential sources.

Split this logic into its own function. It will see future use in the
`s3_client` to assist in self-configuration for required platform specific
configs.