`s3_client`: use `infer_backend_from_uri()` in `self_configure()`
Previously, OCI would mistakenly self-configure itself for the `virtual-host`
style, due to a check with a `ListObjects` request that succeeds.

Oracle states that they only support path-style requests, so this behavior is
strange:

https://www.oracle.com/ca-en/cloud/storage/object-storage/faq/#category-amazon

We now use `infer_backend_from_uri()` to check for the `oracle_s3_compat` backend, and
return `s3_url_style::path` if this backend is detected.