iceberg: add value representation
This adds a representation of record values, for use in Iceberg
integration. Some expected usage of these values are:
- When converting Kafka data records to an intermediary format before
  converting to parquet, that format can be these values, rather than a
  thirdparty representation like Arrow (which may not play well with
  Seastar's allocator, e.g. doesn't use iobufs).
- We can represent manifest entries as these values, and then serialize
  the values to Avro dynamically, rather than using fixed-schema
  codegen. This allows us to include a partition key to the manifest
  entries, which we wouldn't be able to do with the current codegen.

The latter here is my primary motivator, but the implementation is
enough to serve the former use case too.

The structure of the values roughly resembles the Rust Literal
implementation[1], but using variants instead of Rust enums. There is
1:1 correspondence between each value and the types in
iceberg/datatypes.h, though note that there is no strict relationship
between the types and values encoded anywhere (yet?).

This commit includes the definition, an equality operator, a stream
operator, and  a hashing function.

[1] https://github.com/apache/iceberg-rust/blob/main/crates/iceberg/src/spec/values.rs